Improve override_settings tests
I don't think testing codes should be in ``override_settings``
directory because it is not django app but an utility of django
(which mean the user do not add ``'override_settings'`` to them
``INSTALLED_APPS`` for using.) Testing it is only required by
developers or contributers of ``override_settings``

Thus I moved these testing code to ``tests`` directory which is
not installed with ``setup.py``.

And also, I didn't want to install ``override_settings`` for testing
it thus I add codes for adding path to test ``override_settings``
without installing in ``run_tests.py``. Now you can test
``override_settings`` with the commands below::

    $ python setup.py test

    or

    $ python tests/runtest.py

    or even

    $ ./tests/runtest.py